Move server related classes into common package
