--- src/main/java/cz/forgottenempire/arma3servergui/additionalserver/AdditionalServerInstanceInfo.java
@@ -1,16 +1,0 @@
-package cz.forgottenempire.arma3servergui.additionalserver;
-
-import java.time.LocalDateTime;
-import javax.annotation.concurrent.Immutable;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@Immutable
-@AllArgsConstructor
-public class AdditionalServerInstanceInfo {
-    private final long id;
-    private final boolean alive;
-    private final LocalDateTime startedAt;
-    private final Process process;
-}
