--- src/main/java/cz/forgottenempire/arma3servergui/additionalserver/dtos/AdditionalServerDto.java
@@ -1,14 +1,0 @@
-package cz.forgottenempire.arma3servergui.additionalserver.dtos;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-public class AdditionalServerDto {
-    private Long id;
-    private String name;
-    private boolean alive;
-    private String startedAt;
-    private String imageUrl;
-}
