--- src/main/java/cz/forgottenempire/arma3servergui/additionalserver/repositories/AdditionalServerRepository.java
@@ -1,10 +1,0 @@
-package cz.forgottenempire.arma3servergui.additionalserver.repositories;
-
-import cz.forgottenempire.arma3servergui.additionalserver.entities.AdditionalServer;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AdditionalServerRepository extends JpaRepository<AdditionalServer, Long> {
-
-}
