--- src/main/java/cz/forgottenempire/arma3servergui/additionalserver/services/AdditionalServersService.java
@@ -1,19 +1,0 @@
-package cz.forgottenempire.arma3servergui.additionalserver.services;
-
-import cz.forgottenempire.arma3servergui.additionalserver.AdditionalServerInstanceInfo;
-import cz.forgottenempire.arma3servergui.additionalserver.entities.AdditionalServer;
-import java.util.List;
-import java.util.Optional;
-
-public interface AdditionalServersService {
-
-    List<AdditionalServer> getAllServers();
-
-    Optional<AdditionalServer> getServer(Long id);
-
-    AdditionalServerInstanceInfo getServerInstanceInfo(Long id);
-
-    void startServer(Long serverId);
-
-    void stopServer(Long serverId);
-}
