+++ src/main/java/cz/forgottenempire/arma3servergui/common/RestExceptionHandler.java
@@ -2,8 +2,8 @@ package cz.forgottenempire.arma3servergui.common;
 
 import cz.forgottenempire.arma3servergui.common.exceptions.NotFoundException;
 import cz.forgottenempire.arma3servergui.common.exceptions.ServerNotInitializedException;
-import cz.forgottenempire.arma3servergui.server.exceptions.ModifyingRunningServerException;
-import cz.forgottenempire.arma3servergui.server.exceptions.PortAlreadyTakenException;
+import cz.forgottenempire.arma3servergui.server.serverinstance.exceptions.ModifyingRunningServerException;
+import cz.forgottenempire.arma3servergui.server.serverinstance.exceptions.PortAlreadyTakenException;
 import java.util.ArrayList;
 import java.util.List;
 import org.springframework.http.HttpHeaders;
