+++ src/main/java/cz/forgottenempire/arma3servergui/server/additionalserver/dtos/AdditionalServerDto.java
@@ -1,0 +1,14 @@
+package cz.forgottenempire.arma3servergui.server.additionalserver.dtos;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+public class AdditionalServerDto {
+    private Long id;
+    private String name;
+    private boolean alive;
+    private String startedAt;
+    private String imageUrl;
+}
