--- src/main/java/cz/forgottenempire/arma3servergui/server/dtos/ServerDetails.java
@@ -1,21 +1,0 @@
-package cz.forgottenempire.arma3servergui.server.dtos;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class ServerDetails {
-    private long spaceLeft;
-    private long spaceTotal;
-
-    private long memoryLeft;
-    private long memoryTotal;
-
-    private double cpuUsage;
-    private int cpuCount;
-
-    private String osName;
-    private String osVersion;
-    private String osArchitecture;
-}
