--- src/main/java/cz/forgottenempire/arma3servergui/server/dtos/ServerWorkshopModDto.java
@@ -1,9 +1,0 @@
-package cz.forgottenempire.arma3servergui.server.dtos;
-
-import lombok.Data;
-
-@Data
-public class ServerWorkshopModDto {
-    private Long id;
-    private String name;
-}
