--- src/main/java/cz/forgottenempire/arma3servergui/server/exceptions/ModifyingRunningServerException.java
@@ -1,15 +1,0 @@
-package cz.forgottenempire.arma3servergui.server.exceptions;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-public class ModifyingRunningServerException extends RuntimeException{
-
-    public ModifyingRunningServerException() {
-    }
-
-    public ModifyingRunningServerException(String message) {
-        super(message);
-    }
-}
-
