--- src/main/java/cz/forgottenempire/arma3servergui/server/exceptions/PortAlreadyTakenException.java
@@ -1,16 +1,0 @@
-package cz.forgottenempire.arma3servergui.server.exceptions;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(value = HttpStatus.CONFLICT, reason = "port already taken")
-public class PortAlreadyTakenException extends RuntimeException{
-
-    public PortAlreadyTakenException() {
-    }
-
-    public PortAlreadyTakenException(String message) {
-        super(message);
-    }
-}
-
