+++ src/main/java/cz/forgottenempire/arma3servergui/server/serverinstance/dtos/ServerWorkshopModDto.java
@@ -1,0 +1,9 @@
+package cz.forgottenempire.arma3servergui.server.serverinstance.dtos;
+
+import lombok.Data;
+
+@Data
+public class ServerWorkshopModDto {
+    private Long id;
+    private String name;
+}
