+++ src/main/java/cz/forgottenempire/arma3servergui/server/serverinstance/dtos/ServersDto.java
@@ -1,0 +1,13 @@
+package cz.forgottenempire.arma3servergui.server.serverinstance.dtos;
+
+import java.util.List;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class ServersDto {
+    private List<ServerDto> servers;
+}
