+++ src/main/java/cz/forgottenempire/arma3servergui/workshop/services/impl/WorkshopModsServiceImpl.java
@@ -1,7 +1,7 @@
 package cz.forgottenempire.arma3servergui.workshop.services.impl;
 
 import cz.forgottenempire.arma3servergui.common.Constants;
-import cz.forgottenempire.arma3servergui.server.repositories.ServerRepository;
+import cz.forgottenempire.arma3servergui.server.serverinstance.repositories.ServerRepository;
 import cz.forgottenempire.arma3servergui.system.entities.SteamAuth;
 import cz.forgottenempire.arma3servergui.workshop.entities.WorkshopMod;
 import cz.forgottenempire.arma3servergui.workshop.repositories.WorkshopModRepository;
