+++ src/main/resources/templates/serverConfigArma3.ftl
@@ -1,4 +1,4 @@
-<#-- @ftlvariable name="" type="cz.forgottenempire.arma3servergui.server.entities.Server" -->
+<#-- @ftlvariable name="" type="cz.forgottenempire.arma3servergui.server.serverinstance.entities.Server" -->
 
 // GLOBAL SETTINGS
 hostname = "${name}";
